Completed a basic version of the file protocol. Much work needs to be done on server side authentication and data integrity analysis, but for the time being it "does the job" if not particularly nicely.
Server side implementation of the crytographic hashing is required, and
fileservice requires updating to give the actual hash type being used for
the benefit of the server and to erase a potential error state.

I also homogenised the buffer sizes being used to ensure some sort of
consistency, which is nominally located in the framesync.py package.




Signed-off-by: KentJames <jameschristopherkent@gmail.com>